refactor(clean-arch): move buildEvalPrincipalFn/computeComplianceByPrinciple to a shared utils module
useOpenPrinciple.js (compare feature) was reaching into explorer's
components/ folder for domain logic, a cross-feature layering violation.
Both functions now live in utils/evalPrincipal.js; explorerDataHooks.js
re-exports them so ExplorerPage.jsx needs no changes.
