+++ src/quodeq/ui/src/utils/evalPrincipal.js
@@ -1,0 +1,24 @@
+export function computeComplianceByPrinciple(evalData) {
+  const map = new Map();
+  for (const c of (evalData?.compliance || [])) {
+    if (!map.has(c.principle)) map.set(c.principle, []);
+    map.get(c.principle).push(c);
+  }
+  return map;
+}
+
+export function buildEvalPrincipalFn(evalData, complianceByPrinciple, project, runId, dateLabel = '') {
+  const principlesByName = new Map((evalData.principles || []).map((p) => [p.name, p]));
+  const gradesByPrinciple = new Map((evalData.principleGrades || []).map((p) => [p.principle, p]));
+  return function buildEvalPrincipal(principleId) {
+    const principleData = principlesByName.get(principleId);
+    const pg = gradesByPrinciple.get(principleId);
+    return {
+      principle: principleId, score: pg?.score || null, grade: pg?.grade || null,
+      dimension: evalData.dimension || '',
+      project: project || '', runId: runId || '', dateLabel: dateLabel || '',
+      principleData, dimViolations: principleData?.violations || [],
+      dimCompliance: complianceByPrinciple.get(principleId) || [],
+    };
+  };
+}
